Move SimpleTable to separate file
